Reviewed-on: https://git.onlyoffice.com/ONLYOFFICE/onlyoffice-docspace-monday/pulls/3

+++ onlyoffice-monday-gateway/src/main/java/com/onlyoffice/gateway/processor/WebSocketNotificationProcessor.java
@@ -1,3 +1,16 @@
+/**
+ * (c) Copyright Ascensio System SIA 2025
+ *
+ * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License at
+ *
+ * <p>http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * <p>Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.onlyoffice.gateway.processor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -76,7 +89,7 @@ public class WebSocketNotificationProcessor extends TextWebSocketHandler
     }
   }
 
-  private boolean isSessionEligible(WebSocketSession session, int tenantId) {
+  private boolean isSessionEligible(WebSocketSession session, long tenantId) {
     if (!session.isOpen() || session.getUri() == null) return false;
     var token = getSessionToken(session);
     if (token == null) return false;
@@ -89,7 +102,7 @@ public class WebSocketNotificationProcessor extends TextWebSocketHandler
     return (tokenParams != null && !tokenParams.isEmpty()) ? tokenParams.getFirst() : null;
   }
 
-  private boolean verifyTenantId(String token, int tenantId) {
+  private boolean verifyTenantId(String token, long tenantId) {
     try {
       var claims = jwtDecoder.decode(token).getClaims();
       var sessionToken = mapper.convertValue(claims.get("dat"), SessionToken.class);
@@ -100,7 +113,7 @@ public class WebSocketNotificationProcessor extends TextWebSocketHandler
     }
   }
 
-  private void sendNotification(WebSocketSession session, TextMessage message, int tenantId) {
+  private void sendNotification(WebSocketSession session, TextMessage message, long tenantId) {
     try {
       MDC.put("tenant_id", String.valueOf(tenantId));
       MDC.put("session_id", session.getId());
