+++ onlyoffice-monday-tenant/src/main/java/com/onlyoffice/tenant/controller/query/BoardQueryController.java
@@ -1,3 +1,16 @@
+/**
+ * (c) Copyright Ascensio System SIA 2025
+ *
+ * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License at
+ *
+ * <p>http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * <p>Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.onlyoffice.tenant.controller.query;
 
 import com.onlyoffice.common.tenant.transfer.request.query.FindEntity;
@@ -26,7 +39,7 @@ public class BoardQueryController {
 
   @GetMapping("/{boardId}")
   @RateLimiter(name = "findBoard")
-  public ResponseEntity<BoardInformation> findBoard(@PathVariable @Positive int boardId) {
+  public ResponseEntity<BoardInformation> findBoard(@PathVariable @Positive long boardId) {
     return ResponseEntity.ok(queryService.find(FindEntity.builder().id(boardId).build()));
   }
 }
